refactor: PR 피드백 반영 2
1. Sections 일급컬렉션 생성
2. Station 엔티티 equals, hashcode 재정의 제거
3. SectionTest를 일반 테스트 코드로 변경
4. Line - Section 양방향을 고려한 삭제 메서드로 수정
