+++ src/test/java/nextstep/subway/line/domain/SectionTest.java
@@ -2,46 +2,20 @@ package nextstep.subway.line.domain;
 
 import static org.assertj.core.api.Assertions.*;
 
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 
 import nextstep.subway.station.domain.Station;
-import nextstep.subway.station.domain.StationRepository;
 
-@DataJpaTest
 class SectionTest {
-    @Autowired
-    private StationRepository stationRepository;
-
-    @Autowired
-    private LineRepository lineRepository;
-
-    private Station 인천역;
-    private Station 용산역;
-    private Line 일호선;
-    private Section 인천_용산역_구간;
-
-    @AfterEach
-    void tearDown() {
-        stationRepository.deleteAll();
-        lineRepository.deleteAll();
-    }
-
-    @BeforeEach
-    void setUp() {
-        인천역 = stationRepository.save(new Station("인천역"));
-        용산역 = stationRepository.save(new Station("용산역"));
-        일호선 = lineRepository.save(new Line("1호선", "indigo", 인천역, 용산역, 10));
-        인천_용산역_구간 = 일호선.getSections().get(0);
-    }
+    private Station 인천역 = new Station("인천역");
+    private Station 용산역 = new Station("용산역");
+    private Line 일호선 = new Line("1호선", "indigo", 인천역, 용산역, 10);
+    private Section 인천_용산역_구간 = new Section(일호선, 인천역, 용산역, 10);
 
     @Test
     void divideBy() {
         // given
-        final Station 신도림역 = stationRepository.save(new Station("신도림역"));
+        final Station 신도림역 = new Station("신도림역");
         final Section section = new Section(일호선, 인천역, 신도림역, 3);
 
         // when
@@ -55,7 +29,7 @@ class SectionTest {
     @Test
     void isOverlapped() {
         // given
-        final Station 의정부역 = stationRepository.save(new Station("의정부역"));
+        final Station 의정부역 = new Station("의정부역");
         final Section section1 = new Section(일호선, 용산역, 의정부역, 5);
         final Section section2 = new Section(일호선, 인천역, 의정부역, 15);
 
@@ -67,7 +41,7 @@ class SectionTest {
     @Test
     void connectWith() {
         // given
-        final Station 의정부역 = stationRepository.save(new Station("의정부역"));
+        final Station 의정부역 = new Station("의정부역");
         final Section section = new Section(일호선, 용산역, 의정부역, 5);
 
         // when
@@ -82,7 +56,7 @@ class SectionTest {
     @Test
     void isNextSection() {
         // given
-        final Station 의정부역 = stationRepository.save(new Station("의정부역"));
+        final Station 의정부역 = new Station("의정부역");
         final Section section = new Section(일호선, 용산역, 의정부역, 5);
 
         // when, then
